+++ modules/react-components/src/components/banner/index.ts
