+++ modules/react-components/src/components/index.ts
@@ -21,6 +21,7 @@ export * from "./accordion";
 export * from "./alert";
 export * from "./analytics";
 export * from "./avatar";
+export * from "./banner";
 export * from "./brand";
 export * from "./button";
 export * from "./card";
@@ -33,7 +34,6 @@ export * from "./header";
 export * from "./help-panel";
 export * from "./icon";
 export * from "./input";
-export * from "./jumbotron";
 export * from "./language-switcher";
 export * from "./list";
 export * from "./loader";
